refactor: refactor vscode animation, handle empty files and empty prompts
