+++ templates/frontend/src/components/pages/collection-template/vscode-animation/folder-menu/index.js
@@ -1,0 +1,1 @@
+export { default } from './folder-menu';
