+++ templates/frontend/src/components/pages/collection-template/vscode-animation/prompt-input/index.js
@@ -1,0 +1,1 @@
+export { default } from './prompt-input';
