+++ templates/frontend/src/components/pages/collection-template/vscode-animation/template-select/index.js
@@ -1,0 +1,1 @@
+export { default } from './template-select';
