+++ templates/frontend/src/components/pages/collection-template/vscode-animation/template-select/template-select.module.scss
@@ -1,0 +1,80 @@
+.template-select {
+  position: absolute;
+  top: 60px;
+  left: 50%;
+  z-index: 3;
+  transform: translateX(-50%);
+}
+
+.prompt-input {
+  position: absolute;
+  top: 60px;
+  left: 50%;
+  z-index: 3;
+  transform: translateX(-50%);
+}
+
+.select-wrapper {
+  width: 591px;
+  background-color: #252526;
+  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.4);
+}
+
+.select-input {
+  display: flex;
+  align-items: center;
+  height: 29px;
+  padding-left: 4px;
+  margin-top: 7px;
+  margin-right: 6px;
+  margin-bottom: 7px;
+  margin-left: 6px;
+  background: #3c3c3c;
+  border: 1px solid #007acc;
+}
+
+.select-placeholder {
+  color: #ccc;
+  opacity: 0.4;
+}
+
+.select-text {
+  color: #ccc;
+}
+
+.select-options {
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  padding-bottom: 8px;
+  margin-top: 7px;
+}
+
+@keyframes highlightActiveTemplate {
+  0% {
+    background-color: transparent;
+  }
+
+  85% {
+    background-color: transparent;
+  }
+
+  100% {
+    background-color: #062f4a;
+  }
+}
+
+.select-option {
+  width: 100%;
+  padding-top: 4px;
+  padding-right: 12px;
+  padding-bottom: 7px;
+  padding-left: 12px;
+
+  &.active {
+    background-color: #062f4a;
+    animation-name: highlightActiveTemplate;
+    animation-duration: 1s;
+    animation-iteration-count: 1;
+  }
+}
