+++ templates/frontend/src/components/pages/collection-template/vscode-animation/vscode-animation.module.scss
@@ -3,161 +3,3 @@
   width: 100%;
   margin-top: 40px;
 }
-
-// FOLDER MENU STYLES
-.menu {
-  position: absolute;
-  top: 25%;
-  left: 20%;
-  z-index: 3;
-  display: flex;
-  flex-direction: column;
-  width: 293px;
-  padding-top: 5px;
-  padding-bottom: 5px;
-  background: rgba(255, 255, 255, 0.8);
-  border: 0.5px solid rgba(0, 0, 0, 0.12);
-  border-radius: 6px;
-  box-shadow: 0 0 20px rgba(0, 0, 0, 0.15);
-}
-
-@keyframes highlightActiveItem {
-  0% {
-    color: black;
-    background-color: transparent;
-  }
-
-  85% {
-    color: black;
-    background-color: transparent;
-  }
-
-  100% {
-    color: white;
-    background-color: #35f;
-  }
-}
-
-.menu-item {
-  display: flex;
-  justify-content: space-between;
-  padding-top: 3px;
-  padding-right: 10px;
-  padding-bottom: 3px;
-  padding-left: 10px;
-  margin-right: 4px;
-  margin-bottom: 3px;
-  margin-left: 4px;
-  color: black;
-  border-radius: 3px;
-
-  &:last-child {
-    margin-bottom: 0;
-  }
-
-  &.disabled {
-    color: rgba(60, 60, 67, 0.6);
-  }
-
-  &.active {
-    color: white;
-    background-color: #35f;
-    animation-name: highlightActiveItem;
-    animation-duration: 1s;
-    animation-iteration-count: 1;
-  }
-}
-
-.item-shortcut {
-  color: rgba(60, 60, 67, 0.6);
-}
-
-.separator {
-  width: 100%;
-  margin-top: 2px;
-  margin-bottom: 5px;
-  border-bottom: 1px solid rgba(60, 60, 67, 0.18);
-}
-
-// template select
-
-.template-select {
-  position: absolute;
-  top: 60px;
-  left: 50%;
-  z-index: 3;
-  transform: translateX(-50%);
-}
-
-.prompt-input {
-  position: absolute;
-  top: 60px;
-  left: 50%;
-  z-index: 3;
-  transform: translateX(-50%);
-}
-
-.select-wrapper {
-  width: 591px;
-  background-color: #252526;
-  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.4);
-}
-
-.select-input {
-  display: flex;
-  align-items: center;
-  height: 29px;
-  padding-left: 4px;
-  margin-top: 7px;
-  margin-right: 6px;
-  margin-bottom: 7px;
-  margin-left: 6px;
-  background: #3c3c3c;
-  border: 1px solid #007acc;
-}
-
-.select-placeholder {
-  color: #ccc;
-  opacity: 0.4;
-}
-
-.select-text {
-  color: #ccc;
-}
-
-.select-options {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  padding-bottom: 8px;
-  margin-top: 7px;
-}
-
-@keyframes highlightActiveTemplate {
-  0% {
-    background-color: transparent;
-  }
-
-  85% {
-    background-color: transparent;
-  }
-
-  100% {
-    background-color: #062f4a;
-  }
-}
-
-.select-option {
-  width: 100%;
-  padding-top: 4px;
-  padding-right: 12px;
-  padding-bottom: 7px;
-  padding-left: 12px;
-
-  &.active {
-    background-color: #062f4a;
-    animation-name: highlightActiveTemplate;
-    animation-duration: 1s;
-    animation-iteration-count: 1;
-  }
-}
