+++ templates/frontend/src/components/shared/editor/editor.jsx
@@ -27,30 +27,29 @@ const Editor = ({
   readOnly = false,
   content = '',
   language = DEFAULT_LANGUAGE,
-  onFileContentChange,
+  onFileContentChange = () => {},
   theme = 'twilight',
 }) => (
-    <div className={cx('wrapper', className)}>
-      <AceEditor
-        readOnly={readOnly}
-        value={content}
-        mode={language.aceMode.toLowerCase()}
-        theme={theme}
-        name="UNIQUE_ID_OF_DIV"
-        editorProps={{ $blockScrolling: true }}
-        width="100%"
-        height="100%"
-        fontSize="14px"
-        setOptions={{ useWorker: false }}
-        style={{ lineHeight: '22px' }}
-        showPrintMargin={false}
-        onChange={onFileContentChange}
-        onLoad={(editor) => {
-          editor.renderer.setPadding(22);
-          editor.renderer.setScrollMargin(22);
-        }}
-      />
-    </div>
-  );
-
+  <div className={cx('wrapper', className)}>
+    <AceEditor
+      readOnly={readOnly}
+      value={content}
+      mode={language.aceMode.toLowerCase()}
+      theme={theme}
+      name="UNIQUE_ID_OF_DIV"
+      editorProps={{ $blockScrolling: true }}
+      width="100%"
+      height="100%"
+      fontSize="14px"
+      setOptions={{ useWorker: false }}
+      style={{ lineHeight: '22px' }}
+      showPrintMargin={false}
+      onChange={onFileContentChange}
+      onLoad={(editor) => {
+        editor.renderer.setPadding(22);
+        editor.renderer.setScrollMargin(22);
+      }}
+    />
+  </div>
+);
 export default Editor;
